WIP - Replace banner image with vimeo player
This is almost working - just need to work out how to ensure the height
stays 100% somehow when the width of the window is reduced.

Auto-resizing code inspired by this
[here](https://stackoverflow.com/questions/33451481/scale-and-reposition-iframe-like-background-size-cover).
This is the same technique used in the videos below, though we were able
to work around this issue by changing the reflow width which we cannot
do in this case.